% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgroundCorrection.R
\name{backgroundCorrectionNoobFit}
\alias{backgroundCorrectionNoobFit}
\title{Fit Normal and exponential distributions (adapted from \code{SeSAMe})}
\usage{
backgroundCorrectionNoobFit(ib, bg)
}
\arguments{
\item{ib}{Foreground signals.}

\item{bg}{Background signals.}
}
\value{
mu and sigma by fitting background signals with normal distribution.

alpha by fitting foreground signals with exponential distribution.
}
\description{
Fit Normal and exponential distributions (adapted from \code{SeSAMe})
}
