% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes.R
\name{call_genotypes}
\alias{call_genotypes}
\title{Call genotypes based on EM algorithm}
\usage{
call_genotypes(
  RAI,
  pop,
  type,
  maxiter = 50,
  bayesian = FALSE,
  platform = "EPIC",
  verbose = 1
)
}
\arguments{
\item{RAI}{A matrix of RAI (Ratio of Alternative allele Intensity) for probes. Provide probes as rows and samples as columns.}

\item{pop}{Population to be used to extract AFs. One of EAS, AMR, AFR, EUR, SAS, and ALL.}

\item{type}{One of snp_probe, typeI_probe, and typeII_probe.}

\item{maxiter}{Maximal number of iterations for the EM algorithm.}

\item{bayesian}{Use the Bayesian approach to calculate posterior genotype probabilities.}

\item{platform}{EPIC, EPIC_v2, 450K, or MSA.}

\item{verbose}{Verbose mode: 0/1/2.}
}
\value{
A list containing
\item{RAI}{Ratio of Alternative allele Intensity}
\item{shapes}{Shapes of the mixed beta distributions}
\item{weights}{Prior probabilities that the RAI values belong to one of the three genotypes}
\item{U}{Overall probability of RAI values being outlier}
\item{outliers}{Probability of each RAI value being outlier}
\item{logLik}{Log-likelihood}
\item{GP}{Genotype probabilities of the three genotypes}
}
\description{
The Expectation–maximization (EM) algorithm is used to fit a mixture of three beta distributions representing the three genotypes (AA, AB, and BB) and one uniform distribution representing the outliers (adapted from ewastools). Probe-specific weights were used in the EM algorithm.
}
