% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes.R
\name{callGeno}
\alias{callGeno}
\title{Genotype calling (adapted from \code{ewastools})}
\usage{
callGeno(RAI, learn = FALSE, maxiter = 50)
}
\arguments{
\item{RAI}{A matrix of RAI (Ratio of Alternative allele Intensity) for probes. Provide probes as rows and samples as columns.}

\item{learn}{If learn=TRUE, will learn dataset-specific parameters using the EM algorithm. The genotype calling process will be iterated two times, with the first iteration for selecting probes and the second iteration for re-training.}

\item{maxiter}{Maximal number of iterations of the EM algorithm learning the mixture model.}
}
\value{
A list containing
\item{par}{Parameters of the mixture model}
\item{loglik}{Log-likelihood in each iteration of the EM algorithm}
\item{outliers}{A-posteriori probability of SNP being an outlier}
\item{gamma}{A-posteriori probabilities for each of the three genotypes}
}
\description{
A mixture model (3 Beta distributions, 1 uniform distribution for outliers) is fitted to probes.
After learning the model parameters via Expectation-Maximization (EM) algorithm, the probability of being an outlier is computed for each SNP.
}
