% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping_probe.R
\name{callGeno_genotyping}
\alias{callGeno_genotyping}
\title{Call genotypes for genotyping probes}
\usage{
callGeno_genotyping(rgData, plotBeta = FALSE, vcf = FALSE)
}
\arguments{
\item{rgData}{Noob and dye-bias corrected signals produced by using \code{correct_noob_dye}.}

\item{plotBeta}{If TRUE, plot beta distributions for reference homozygous, heterozygous, and alternative homozygous.}

\item{vcf}{If TRUE, will write a VCF file in the current directory.}
}
\value{
A matrix of genotype calls.
}
\description{
Call genotypes for genotyping probes
}
