% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeI_probe.R
\name{callGeno_typeI}
\alias{callGeno_typeI}
\title{Call genotypes for Type I probes}
\usage{
callGeno_typeI(
  rgData,
  plotRAI = FALSE,
  vcf = FALSE,
  vcfName = "genotypes.typeI_probe.vcf",
  bw = 0.04,
  minDens = 0.001,
  GP_cutoff = 0.9,
  outlier_cutoff = "max",
  missing_cutoff = 0.1,
  R2_cutoff_up = 1.1,
  R2_cutoff_down = 0.75,
  MAF_cutoff = 0.01,
  HWE_cutoff = 1e-06,
  cpu = 1,
  pop = "EAS",
  bayesian = FALSE,
  platform = "EPIC",
  verbose = 1
)
}
\arguments{
\item{rgData}{Noob and dye-bias corrected signals produced by using \code{correct_noob_dye}.}

\item{plotRAI}{If TRUE, plot distribution of RAIs.}

\item{vcf}{If TRUE, will write a VCF file in the current directory.}

\item{vcfName}{VCF file name. Only effective when vcf=TRUE.}

\item{bw}{band width.}

\item{minDens}{A parameter for mode test. Minimum density for a valid peak.}

\item{GP_cutoff}{When calculating missing rate, genotypes with the highest genotype probability < GP_cutoff will be treated as missing.}

\item{outlier_cutoff}{"max" or a number ranging from 0 to 1. If outlier_cutoff="max", genotypes with outlier probability larger than all of the three genotype probabilities will be set as missing. If outlier_cutoff is a number, genotypes with outlier probability > outlier_cutoff will be set as missing.}

\item{missing_cutoff}{Missing rate cutoff to filter variants. Note that for VCF output, variants with missing rate above the cutoff will be marked in the \code{FILTER} column. For the returned dosage matrix, variants with missing rate above the cutoff will be removed.}

\item{R2_cutoff_up, R2_cutoff_down}{R-square cutoffs to filter variants (Variants with R-square > R2_cutoff_up or < R2_cutoff_down should be removed). Note that for VCF output, variants with R-square outside this range will be marked in the \code{FILTER} column. For the returned dosage matrix, variants with R-square outside this range will be removed.}

\item{MAF_cutoff}{A MAF cutoff to filter variants. Note that for VCF output, variants with MAF below the cutoff will be marked in the \code{FILTER} column. For the returned dosage matrix, variants with MAF below the cutoff will be removed.}

\item{HWE_cutoff}{HWE p value cutoff to filter variants. Note that for VCF output, variants with HWE p value below the cutoff will be marked in the \code{FILTER} column. For the returned dosage matrix, variants with HWE p value below the cutoff will be removed.}

\item{cpu}{Number of CPU cores.}

\item{pop}{Population. One of EAS, AMR, AFR, EUR, SAS, and ALL. Only probes with MAF of matching population > 0.01 will be kept.}

\item{bayesian}{Use the Bayesian approach to calculate posterior genotype probabilities.}

\item{platform}{EPIC, 450K, or MSA.}

\item{verbose}{Verbose mode: 0/1/2.}
}
\value{
A list containing
\item{dosage}{A matrix of genotype calls. Variants with R2s, HWE p values, MAFs, or missing rates beyond the cutoffs are removed.}
\item{genotypes}{A list containing RAI, shapes of the mixed beta distributions, prior probabilities that the RAI values belong to one of the three genotypes, proportion of RAI values being outlier (U), and genotype probability (GP).}
}
\description{
Call genotypes for Type I probes
}
