% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R
\name{constrain_R2}
\alias{constrain_R2}
\title{Constrain R2}
\usage{
constrain_R2(R2)
}
\arguments{
\item{R2}{R-square}
}
\value{
Constrained R2
}
\description{
R2 is calculated by var(G)/2p(1-p), where G is dosage genotype and p is allele frequency. Variants with 1<R2<=1.1 are constrained to 1. Variants with R2>1.1 (marked as .) are recommended to remove.
}
