% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{control_probes_MSA}
\alias{control_probes_MSA}
\title{Control probes for MSA}
\format{
A data frame with 2511 rows and 4 columns:
\describe{
\item{Address}{Address}
\item{Type}{Type}
\item{Color}{Color}
\item{ExtendedType}{ExtendedType}
}
}
\usage{
data(control_probes_MSA)
}
\description{
Control probes for IlluminaHumanMethylationMSA.
}
\keyword{datasets}
