% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgroundCorrection.R
\name{correct_noob_dye}
\alias{correct_noob_dye}
\title{Noob and dye-bias correction}
\usage{
correct_noob_dye(target, platform = "EPIC", cpu = 1)
}
\arguments{
\item{target}{A data frame of two columns: Sample_Name, Basename, where Basename tells the location of IDAT files.}

\item{platform}{EPIC, EPIC_v2, 450K, or MSA.}

\item{cpu}{Number of CPU.}
}
\value{
A list of noob and dye-bias corrected signals containing:
\item{AR}{ - A matrix of probeA signals in Red channel}
\item{AG}{ - A matrix of probeA signals in Green channel}
\item{BR}{ - A matrix of probeB signals in Red channel}
\item{BG}{ - A matrix of probeB signals in Green channel}
}
\description{
Noob and dye-bias correction
}
