% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R
\name{dosage2hard}
\alias{dosage2hard}
\title{Get hard genotypes from genotype probabilities}
\usage{
dosage2hard(AA, AB, BB)
}
\arguments{
\item{AA}{Genotype probability of AA or 0/0.}

\item{AB}{Genotype probability of AB or 0/1.}

\item{BB}{Genotype probability of BB or 1/1.}
}
\value{
A matrix of hard genotypes.
}
\description{
Hard genotype is defined as the genotype with highest genotype probability.
}
