% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes_bayesian.R
\name{e_step}
\alias{e_step}
\title{The Expectation–maximization (EM) algorithm: E step}
\usage{
e_step(assignments, fits)
}
\arguments{
\item{assignments}{A data frame containing:
\item{Probe}{Probe ID}
\item{Sample}{Sample ID}
\item{RAI}{Ratio of Alternative allele Intensity}
\item{Cluster}{Cluster IDs}}

\item{fits}{A list produced in m_step}
}
\value{
A data frame with re-assigned cluster IDs
}
\description{
The Expectation–maximization (EM) algorithm: E step
}
