% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes_bayesian.R,
%   R/call_genotypes_deprecated.R
\name{eBeta}
\alias{eBeta}
\title{Moments estimator for beta distribution (adapted from ewastools)}
\usage{
eBeta(x, w)

eBeta(x, w)
}
\description{
Moments estimator for beta distribution (adapted from ewastools)

eBeta (adapted from \code{ewastools})
}
