% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R,
%   R/data_representation_deprecated.R
\name{filter_by_AF}
\alias{filter_by_AF}
\title{Filter by AF}
\usage{
filter_by_AF(AF, MAF_cutoff = 0.01)

filter_by_AF(AF, MAF_cutoff = 0.01)
}
\arguments{
\item{AF}{Allele frequency}

\item{MAF_cutoff}{An MAF (Minor allele frequency) cutoff to filter variants.}
}
\value{
Whether the variant passed filtering.

Whether the variant passed filtering.
}
\description{
Variants with MAF<0.01 are recommended to remove.

Variants with MAF<0.01 are recommended to remove.
}
