% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R
\name{filter_by_HWE}
\alias{filter_by_HWE}
\title{Filter by Hardy–Weinberg Equilibrium (HWE) p value}
\usage{
filter_by_HWE(hwe_p, HWE_cutoff = 1e-06)
}
\arguments{
\item{hwe_p}{Hardy–Weinberg Equilibrium (HWE) p values}

\item{HWE_cutoff}{A HWE p value cutoff to filter variants.}
}
\value{
Whether the variant passed filtering.
}
\description{
Variants with Hardy–Weinberg Equilibrium (HWE) p value < HWE_cutoff are recommended to remove.
}
