% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R
\name{filter_by_missing}
\alias{filter_by_missing}
\title{Filter by missing rate}
\usage{
filter_by_missing(missingrate, missing_cutoff = 0.1)
}
\arguments{
\item{missingrate}{Missing rates}

\item{missing_cutoff}{Missing rate cutoff to filter variants.}
}
\value{
Whether the variant passed filtering.
}
\description{
Variants with missing rate > missing_cutoff are recommended to remove.
}
