% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R
\name{filter_by_R2}
\alias{filter_by_R2}
\title{Filter by R2}
\usage{
filter_by_R2(R2)
}
\arguments{
\item{R2}{R-square}
}
\value{
Whether the variant passed filtering.
}
\description{
Variants with R2>1.1 (marked as .) or R2<0.7 are recommended to remove.
}
