% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDetection.R
\name{findCentralFromTwoPeaks}
\alias{findCentralFromTwoPeaks}
\title{Find the central peak from a distribution of two peaks}
\usage{
findCentralFromTwoPeaks(locations)
}
\arguments{
\item{locations}{A vector of two, indicating the locations of the two peaks.}
}
\value{
A vector of three, with the central one being the location of the central peak.
}
\description{
Find the central peak from a distribution of two peaks
}
