% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes.R
\name{fit_beta_em}
\alias{fit_beta_em}
\title{Estimate mixed beta distribution parameters based on EM algorithm}
\usage{
fit_beta_em(RAI, maxiter = 50, verbose = 1)
}
\arguments{
\item{RAI}{A matrix of RAI (Ratio of Alternative allele Intensity) for probes. Provide probes as rows and samples as columns.}

\item{maxiter}{Maximal number of iterations for the EM algorithm.}

\item{verbose}{Verbose mode: 0/1/2.}
}
\value{
A list containing
\item{shapes}{Shapes of the mixed beta distributions}
\item{weights}{Prior probabilities that the RAI values belong to one of the three genotypes}
\item{U}{Overall probability of RAI values being outlier}
\item{outliers}{Probability of each RAI value being outlier}
\item{logLik}{Log-likelihood}
\item{GP}{Genotype probabilities of the three genotypes}
}
\description{
The Expectation–maximization (EM) algorithm is used to fit a mixture of three beta distributions representing the three genotypes (AA, AB, and BB) and one uniform distribution representing the outliers (adapted from ewastools).
}
