% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes_bayesian.R
\name{fit_beta_mle}
\alias{fit_beta_mle}
\title{Fit beta distribution based on Maximum likelihood estimation}
\usage{
fit_beta_mle(x)
}
\arguments{
\item{x}{A vector of RAI values.}
}
\value{
A data frame with one row and four columns:
\item{shape1}{The first parameter for beta distribution}
\item{shape2}{The second parameter for beta distribution}
\item{mLL}{Minus log-likelihood scaled by number of data points}
\item{number}{The length of x}
}
\description{
Fit beta distribution based on Maximum likelihood estimation
}
