% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R
\name{format_genotypes}
\alias{format_genotypes}
\title{Format genotype calls}
\usage{
format_genotypes(
  genotypes,
  vcf = FALSE,
  vcfName,
  GP_cutoff = 0.9,
  outlier_cutoff = "max",
  missing_cutoff = 0.1,
  R2_cutoff_up = 1.1,
  R2_cutoff_down = 0.75,
  MAF_cutoff = 0.01,
  HWE_cutoff = 1e-06,
  pop = "ALL",
  type,
  plotAF = FALSE,
  platform = "EPIC"
)
}
\arguments{
\item{genotypes}{Genotype calls.}

\item{vcf}{If TRUE, will write a VCF file in the current directory.}

\item{vcfName}{VCF file name. Only effective when vcf=TRUE.}

\item{GP_cutoff}{When calculating missing rate, genotypes with the highest genotype probability < GP_cutoff will be treated as missing.}

\item{outlier_cutoff}{"max" or a number ranging from 0 to 1. If outlier_cutoff="max", genotypes with outlier probability larger than all of the three genotype probabilities will be set as missing. If outlier_cutoff is a number, genotypes with outlier probability > outlier_cutoff will be set as missing.}

\item{missing_cutoff}{Missing rate cutoff to filter variants. Note that for VCF output, variants with missing rate above the cutoff will be marked in the \code{FILTER} column. For the returned dosage matrix, variants with missing rate above the cutoff will be removed.}

\item{R2_cutoff_up, R2_cutoff_down}{R-square cutoffs to filter variants (Variants with R-square > R2_cutoff_up or < R2_cutoff_down should be removed). Note that for VCF output, variants with R-square outside this range will be marked in the \code{FILTER} column. For the returned dosage matrix, variants with R-square outside this range will be removed.}

\item{MAF_cutoff}{MAF cutoff to filter variants. Note that for VCF output, variants with MAF below the cutoff will be marked in the \code{FILTER} column. For the returned dosage matrix, variants with MAF below the cutoff will be removed.}

\item{HWE_cutoff}{HWE p value cutoff to filter variants. Note that for VCF output, variants with HWE p value below the cutoff will be marked in the \code{FILTER} column. For the returned dosage matrix, variants with HWE p value below the cutoff will be removed.}

\item{pop}{Population to be used to extract AFs. One of EAS, AMR, AFR, EUR, SAS, and ALL.}

\item{type}{One of snp_probe, typeI_probe, and typeII_probe.}

\item{plotAF}{To plot the distribution of AFs in 1KGP and input data.}

\item{platform}{EPIC, EPIC_v2, 450K, or MSA.}
}
\value{
A matrix of genotype calls. Variants with R2s, HWE p values, MAFs, or missing rates beyond the cutoffs are removed.
}
\description{
Format genotype calls
}
