% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes.R
\name{get_AF}
\alias{get_AF}
\title{Extract AFs from matching population in the 1000 Genomes Project (1KGP)}
\usage{
get_AF(pop = "EAS", type, platform = "EPIC")
}
\arguments{
\item{pop}{Population to be used to extract AFs. One of EAS, AMR, AFR, EUR, SAS, and ALL.}

\item{type}{One of snp_probe, typeI_probe, and typeII_probe.}

\item{platform}{EPIC, EPIC_v2, 450K, or MSA.}
}
\value{
A vector of AFs
}
\description{
Extract AFs from matching population in the 1000 Genomes Project (1KGP)
}
