% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes_bayesian.R
\name{get_GP}
\alias{get_GP}
\title{Infer posterior genotype probabilities based on the Bayesian approach}
\usage{
get_GP(RAI, shapes, probe2af)
}
\arguments{
\item{RAI}{A matrix of RAI (Ratio of Alternative allele Intensity) for probes. Provide probes as rows and samples as columns.}

\item{shapes}{A data frame (3x2) containing the two shapes for beta distributions of the three clusters.}

\item{probe2af}{A vector of AFs extracted from matching population in 1KGP.}
}
\value{
A list containing
\item{pAA}{Posterior genotype probability of AA}
\item{pAB}{Posterior genotype probability of AB}
\item{pBB}{Posterior genotype probability of BB}
}
\description{
Prior genotype probabilities were inferred from AFs of matching population in the 1000 Genomes Project (1KGP).
}
