% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes_bayesian.R
\name{get_GQ}
\alias{get_GQ}
\title{Calculate genotype quality (GQ)}
\usage{
get_GQ(PL)
}
\arguments{
\item{A}{list of Phred-scaled genotype likelihoods containing
\item{AA}{PL(AA)}
\item{AB}{PL(AB)}
\item{BB}{PL(BB)}}
}
\value{
A matrix of genotype qualities
}
\description{
Calculate genotype quality (GQ)
}
