% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrate_genotypes.R
\name{get_indAF}
\alias{get_indAF}
\title{Calculate individual-specific AFs}
\usage{
get_indAF(snpvec, refPC, studyPC)
}
\arguments{
\item{snpvec}{A vector of SNP IDs.}

\item{refPC}{Top PCs in the reference.}

\item{studyPC}{Top PCs in study samples.}
}
\value{
A matrix of individual-specific AFs.
}
\description{
Calculate individual-specific AFs
}
