% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes_bayesian.R
\name{get_PL}
\alias{get_PL}
\title{Calculate Phred-scaled genotype likelihood (PL)}
\usage{
get_PL(GP)
}
\arguments{
\item{GP}{A list of Genotype probabilities containing
\item{pAA}{Posterior genotype probability of AA}
\item{pAB}{Posterior genotype probability of AB}
\item{pBB}{Posterior genotype probability of BB}}
}
\value{
A list of Phred-scaled genotype likelihoods containing
\item{AA}{PL(AA)}
\item{AB}{PL(AB)}
\item{BB}{PL(BB)}
}
\description{
Calculate Phred-scaled genotype likelihood (PL)
}
