% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{get_target}
\alias{get_target}
\title{Get example IDAT file list}
\usage{
get_target(platform = "EPIC")
}
\arguments{
\item{platform}{One of "EPIC" and "450K"}
}
\value{
A data frame of the IDAT file list
}
\description{
Get example IDAT file list
}
