% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs_snp_probe.R
\name{getAB_ccs}
\alias{getAB_ccs}
\title{Get allelic balances for CCS SNP probes}
\usage{
getAB_ccs(rgData)
}
\arguments{
\item{rgData}{Noob and dye-bias corrected signals produced by using \code{correct_noob_dye}.}
}
\value{
Allelic balances.
}
\description{
Get allelic balances for CCS SNP probes
}
