% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotyping_probe.R
\name{getAB_genotyping}
\alias{getAB_genotyping}
\title{Get allelic balances for genotyping probes}
\usage{
getAB_genotyping(rgData)
}
\arguments{
\item{rgData}{Noob and dye-bias corrected signals produced by using \code{correct_noob_dye}.}
}
\value{
Allelic balances.
}
\description{
Get allelic balances for genotyping probes
}
