% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeII_snp_probe.R
\name{getAB_typeII}
\alias{getAB_typeII}
\title{Get allelic balances for Type-II SNP probes}
\usage{
getAB_typeII(rgData, pop = "ALL")
}
\arguments{
\item{rgData}{Noob and dye-bias corrected signals produced by using \code{correct_noob_dye}.}

\item{pop}{Population. One of EAS, AMR, AFR, EUR, SAS, and ALL. Only probes with MAF of matching population > 0.01 will be kept. Only effective when train=TRUE.}
}
\value{
Allelic balances.
}
\description{
Get allelic balances for Type-II SNP probes
}
