% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_genotype_calling.R
\name{getInbreed}
\alias{getInbreed}
\title{Calculate inbreeding coefficients according to Genotypes}
\usage{
getInbreed(dosage)
}
\arguments{
\item{genotype}{Genotype matrix, with each row indicates a SNP and each column indicates a sample. Dosage genotypes are forced to binary (0, 1, or 2) by using the \code{round} function.}
}
\value{
A data frame of inbreeding coefficients and sample contamination status.
}
\description{
Calculate inbreeding coefficients according to Genotypes
}
