% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDetection.R
\name{getMod}
\alias{getMod}
\title{Estimate mode location for each probe}
\usage{
getMod(x, minDens = 0.01, cpu = 1)
}
\arguments{
\item{x}{Matrix of beta or RAI values. Row names must be supplied.}

\item{minDens}{Minimum density for a valid peak.}

\item{cpu}{Number of CPU.}
}
\value{
A data frame of mode locations.
}
\description{
Estimate mode location for each probe
}
