% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeII_probe.R
\name{getRAI_typeII}
\alias{getRAI_typeII}
\title{Get RAI (Ratio of Alternative allele Intensity) for Type-II probes}
\usage{
getRAI_typeII(rgData, pop = "ALL")
}
\arguments{
\item{rgData}{Noob and dye-bias corrected signals produced by using \code{correct_noob_dye}.}

\item{pop}{Population. One of EAS, AMR, AFR, EUR, SAS, and ALL. Only probes with MAF of matching population > 0.01 will be kept. Only effective when train=TRUE.}
}
\value{
RAI (Ratio of Alternative allele Intensity).
}
\description{
Get RAI (Ratio of Alternative allele Intensity) for Type-II probes
}
