% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_genotype_calling.R
\name{getRelation}
\alias{getRelation}
\title{Get sample relationships}
\usage{
getRelation(phi)
}
\arguments{
\item{Phi}{A vector of kinship coefficient (Phi).}
}
\value{
A vector of sample relationships.
}
\description{
Get sample relationships
}
