% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EM_deprecated.R
\name{iterate_em}
\alias{iterate_em}
\title{Get clusters based on the Expectation–maximization (EM) algorithm}
\usage{
iterate_em(assignments)
}
\arguments{
\item{assignments}{A data frame containing:
\item{Probe}{Probe ID}
\item{Sample}{Sample ID}
\item{RAI}{Ratio of Alternative allele Intensity}
\item{Cluster}{Randomly assigned three clusters indicating the three genotypes}}
}
\value{
A list of two elements:
\item{assignments}{A data frame with re-assigned clusters}
\item{fits}{A data frame containing the two shape parameters for the three re-assigned clusters}
}
\description{
Get clusters based on the Expectation–maximization (EM) algorithm
}
