% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes_bayesian.R
\name{m_step}
\alias{m_step}
\title{The Expectation–maximization (EM) algorithm: M step}
\usage{
m_step(assignments)
}
\arguments{
\item{assignments}{A data frame containing:
\item{Probe}{Probe ID}
\item{Sample}{Sample ID}
\item{RAI}{Ratio of Alternative allele Intensity}
\item{Cluster}{Cluster IDs}}
}
\value{
A list of two elements:
\item{parameters}{A data frame of: 1) cluster ID, 2) shape1, 3) shape2, 4) probability of being an outlier, and 5) prior probabilitiy of the genotype being AA,AB,or BB}
\item{mLL_norm}{Minus log-likelihood scaled by number of data points}
}
\description{
The Expectation–maximization (EM) algorithm: M step
}
