% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnfst}
\alias{mnfst}
\title{EPIC manifest file}
\format{
A data frame with 866554 rows and 5 columns:
\describe{
\item{Name}
\item{AddressA_ID}
\item{AddressB_ID}
\item{Infinium_Design_Type}
\item{Color_Channel}
}
}
\source{
\url{https://webdata.illumina.com/downloads/productfiles/methylationEPIC/infinium-methylationepic-v-1-0-b5-manifest-file-csv.zip}
}
\usage{
data(mnfst)
}
\description{
A dataset containing all EPIC probes information.
}
\keyword{datasets}
