% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnfst_450K}
\alias{mnfst_450K}
\title{450K manifest file}
\format{
A data frame with 486428 rows and 5 columns:
\describe{
\item{Name}
\item{AddressA_ID}
\item{AddressB_ID}
\item{Infinium_Design_Type}
\item{Color_Channel}
}
}
\source{
\url{https://webdata.illumina.com/downloads/productfiles/humanmethylation450/humanmethylation450_15017482_v1-2.csv}
}
\usage{
data(mnfst_450K)
}
\description{
A dataset containing all 450K probes information.
}
\keyword{datasets}
