% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mnfst_MSA}
\alias{mnfst_MSA}
\title{MSA manifest file}
\format{
A data frame with 269110 rows and 5 columns:
\describe{
\item{Name}{CpG name}
\item{AddressA_ID}{AdressA ID}
\item{AddressB_ID}{AdressB ID}
\item{Infinium_Design_Type}{Infinium design type}
\item{Color_Channel}{Color channel}
}
}
\source{
\url{https://support.illumina.com/array/array_kits/infinium-methylation-screening-array.html}
}
\usage{
data(mnfst_MSA)
}
\description{
A dataset containing all MSA probes information.
}
\keyword{datasets}
