% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_probe.R
\name{mval2beta}
\alias{mval2beta}
\title{Convert M values to beta values}
\usage{
mval2beta(mval)
}
\arguments{
\item{mval}{M value matrix.}
}
\value{
Beta value matrix.
}
\description{
Convert M values to beta values
}
