% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgroundCorrection.R
\name{normExpSignal}
\alias{normExpSignal}
\title{Normal-exponential deconvolution (adapted from \code{SeSAMe})}
\usage{
normExpSignal(mu, sigma, alpha, x)
}
\arguments{
\item{mu, sigma}{Background signal parameters returned by \code{backgroundCorrectionNoobFit}.}

\item{alpha}{Foreground signal parameters returned by \code{backgroundCorrectionNoobFit}.}

\item{x}{Foreground signals to be corrected.}
}
\value{
The conditional expectation of the signal given the observed foreground and background.
}
\description{
Normal-exponential deconvolution (adapted from \code{SeSAMe})
}
