% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrate_genotypes.R
\name{plot_PCA}
\alias{plot_PCA}
\title{To plot the projection of study samples in reference ancestry space}
\usage{
plot_PCA(refPC, studyPC)
}
\arguments{
\item{refPC}{Top PCs in the reference}

\item{studyPC}{Top PCs in study samples}
}
\description{
To plot the projection of study samples in reference ancestry space
}
