% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R
\name{plotAF_func}
\alias{plotAF_func}
\title{Plot the distribution of AFs in 1KGP and input data.}
\usage{
plotAF_func(AF_input, AF_1KGP, pop, type)
}
\arguments{
\item{AF_input}{A vector.}

\item{AF_1KGP}{A vector.}

\item{pop}{Population. One of EAS, AMR, AFR, EUR, SAS, and ALL.}

\item{type}{One of snp_probe, typeI_ccs_probe, and typeII_ccs_probe.}
}
\description{
Plot the distribution of AFs in 1KGP and input data.
}
