% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes_bayesian.R
\name{plotIter_func}
\alias{plotIter_func}
\title{Plot distribution of RAIs for each iteration}
\usage{
plotIter_func(iterations, type)
}
\arguments{
\item{type}{One of snp_probe, typeI_ccs_probe, and typeII_ccs_probe.}
}
\description{
Plot distribution of RAIs for each iteration
}
