% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_individual_AF.R
\name{plotPCA}
\alias{plotPCA}
\title{To plot the projection of study samples in reference ancestry space}
\usage{
plotPCA(refPC, studyPC)
}
\arguments{
\item{refPC}{Top PCs in the reference}

\item{studyPC}{Top PCs in study samples}
}
\description{
To plot the projection of study samples in reference ancestry space
}
