% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{pprocrustes}
\alias{pprocrustes}
\title{Projection Procrustes Analysis}
\usage{
pprocrustes(
  refPC_new,
  refPC,
  MAX_ITER = 10000,
  THRESHOLD = 1e-06,
  PROCRUSTES_SCALE = 0
)
}
\arguments{
\item{refPC_new}{Top PCs in the combination of reference samples and one study sample}

\item{refPC}{Top PCs in the reference samples}

\item{MAX_ITER}{Maximum iterations for the projection Procrustes analysis}

\item{THRESHOLD}{Convergence criterion for the projection Procrustes analysis}

\item{PROCRUSTES_SCALE}{Fit the scaling parameter to maximize similarity}
}
\value{
Projection Procrustes Analysis results
}
\description{
Adapted from http://csg.sph.umich.edu/chaolong/LASER
}
