% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_genotype_calling.R
\name{predContamination}
\alias{predContamination}
\title{Predict sample contamination according to inbreeding coefficients}
\usage{
predContamination(x)
}
\arguments{
\item{F}{A vector of inbreeding coefficients.}
}
\value{
A vector of whether samples are contaminated (outlier) or not contaminated (normal).
}
\description{
Samples with inbreeding coefficients beyond 3 SDs of the mean are considered as contaminated.
}
