% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probeInfo_typeII}
\alias{probeInfo_typeII}
\title{Type-II SNP probe information}
\format{
A data frame with 4872 rows and 8 columns:
\describe{
\item{Chr}{Chromosome ID}
\item{Pos}{Position}
\item{SNP}{SNP ID targeted by the CpG}
\item{RefAllele}{Reference allele}
\item{AltAllele}{Alternative allele}
\item{CpG}{CpG}
\item{Color}{Color channel}
\item{Group}{Probe types, color channel, and signal corresponds to alternative allele}
}
}
\source{
\url{https://webdata.illumina.com/downloads/productfiles/methylationEPIC/infinium-methylationepic-v-1-0-b5-manifest-file-csv.zip}
}
\usage{
data(probeInfo_ccs)
}
\description{
A dataset containing Type-II SNP probe information.
}
\keyword{datasets}
