% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probelist_450K}
\alias{probelist_450K}
\title{Probe list for 450K}
\format{
A data frame with 12644 rows and 2 columns:
\describe{
\item{CpG}{CpG list}
\item{Type}{Probe types}
}
}
\usage{
data(probelist_450K)
}
\description{
A dataset containing the list of 53 SNP probes on autosomes, 712 Type I probes, and 11875 type II probes.
}
\keyword{datasets}
