% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probelist_MSA}
\alias{probelist_MSA}
\title{Probe list for MSA}
\format{
A data frame with 12926 rows and 3 columns:
\describe{
\item{CpG}{CpG list}
\item{Type}{Probe types}
\item{A2}{Alternative alleles}
}
}
\usage{
data(probelist_MSA)
}
\description{
A dataset containing the list of SNP probes, Type I probes, and type II probes.
}
\keyword{datasets}
