% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trace.R
\name{procrustes}
\alias{procrustes}
\title{Standard Procrustes Analysis}
\usage{
procrustes(refPC_new, refPC, PROCRUSTES_SCALE = 0)
}
\arguments{
\item{refPC_new}{Top PCs in the combination of reference samples and one study sample}

\item{refPC}{Top PCs in the reference samples}

\item{PROCRUSTES_SCALE}{Fit the scaling parameter to maximize similarity}
}
\value{
Procrustes Analysis results
}
\description{
Adapted from http://csg.sph.umich.edu/chaolong/LASER
}
