% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refGeno_1KGP3_SNP_failQC}
\alias{refGeno_1KGP3_SNP_failQC}
\title{SNPs in 1KGP with HWE<1e-20 or F_MISSING>=0.05}
\format{
A vector with 491 items:
\describe{
\item{Value}{SNP rs ID}
}
}
\usage{
data(refGeno_1KGP3_SNP_failQC)
}
\description{
SNPs to be removed in TRACE PCA
}
\keyword{datasets}
