% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDetection.R
\name{rmSmallerFromClosestPeaks}
\alias{rmSmallerFromClosestPeaks}
\title{Find the closest peaks and remove the smaller one}
\usage{
rmSmallerFromClosestPeaks(modes)
}
\arguments{
\item{modes}{A matrix of two columns (col names: loc, dens) and at least two rows.}
}
\value{
The same matrix with one row deleted.
}
\description{
Find the closest peaks and remove the smaller one
}
