% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sam2pop}
\alias{sam2pop}
\title{Population information for the 1KGP samples}
\format{
A vector with 2504 items:
\describe{
\item{Name}{Sample ID}
\item{Value}{Population}
}
}
\usage{
data(sam2pop)
}
\description{
A vector of population information for the 2504 samples in 1KGP.
}
\keyword{datasets}
