% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes_deprecated.R
\name{callGeno}
\alias{callGeno}
\title{Call genotypes (adapted from \code{ewastools})}
\usage{
callGeno(RAI, learn = FALSE, maxiter = 50)
}
\arguments{
\item{RAI}{A matrix of RAI (Ratio of Alternative allele Intensity) for probes. Provide probes as rows and samples as columns.}

\item{learn}{If learn=TRUE, will learn dataset-specific parameters using the EM algorithm. The genotype calling process will be iterated two times, with the first iteration for selecting probes and the second iteration for re-training.}

\item{maxiter}{Maximal number of iterations of the EM algorithm learning the mixture model.}
}
\value{
A list containing
\item{par}{Parameters of the mixture model}
\item{loglik}{Log-likelihood in each iteration of the EM algorithm}
\item{outliers}{A-posteriori probability of SNP being an outlier}
\item{gamma}{A-posteriori probabilities for each of the three genotypes}
}
\description{
A mixture model (3 Beta distributions, 1 uniform distribution for outliers) is fitted to probes. The model parameters are learnt via Expectation-Maximization (EM) algorithm. We used a "two-step" strategy to call genotypes, which is run by fitting the mixture model twice. For the first step, we fit the mixture model, call genotypes, and filter out the probes deviating Hardy-Weinberg Equilibrium (HWE, P < 1E-6). For the second step, we fit the model and call genotypes again, using the pre-filtered data.
}
