% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/snp_probe.R
\name{callGeno_snp}
\alias{callGeno_snp}
\title{Call genotypes for SNP probes}
\usage{
callGeno_snp(
  rgData,
  input = "raw",
  plotBeta = FALSE,
  vcf = FALSE,
  R2_cutoff_up = 1.1,
  R2_cutoff_down = 0.7,
  MAF_cutoff = 0.01
)
}
\arguments{
\item{rgData}{Noob and dye-bias corrected signals produced by using \code{correct_noob_dye}.}

\item{input}{Input data types. One of "raw", "beta", and "mval". If input is "beta" or "mval", please use probes as rows and samples as columns.}

\item{plotBeta}{If TRUE, plot beta distributions for reference homozygous, heterozygous, and alternative homozygous.}

\item{vcf}{If TRUE, will write a VCF file in the current directory.}

\item{R2_cutoff_up, R2_cutoff_down}{R-square cutoffs to filter variants (Variants with R-square > R2_cutoff_up or < R2_cutoff_down should be removed). Note that for VCF output, variants with R-square outside this range will be marked in the \code{FILTER} column. For the returned matrix, variants with R-square outside this range will be removed.}

\item{MAF_cutoff}{A MAF cutoff to filter variants. Note that for VCF output, variants with MAF below the cutoff will be marked in the \code{FILTER} column. For the returned matrix, variants with MAF below the cutoff will be removed.}
}
\value{
A matrix of genotype calls.
}
\description{
Call genotypes for SNP probes
}
