% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeI_ccs_probe.R
\name{callGeno_typeI}
\alias{callGeno_typeI}
\title{Call genotypes for Type I CCS probes}
\usage{
callGeno_typeI(
  rgData,
  plotBeta = FALSE,
  vcf = FALSE,
  vcfName = "genotypes.typeI_ccs_probe.vcf",
  R2_cutoff_up = 1.1,
  R2_cutoff_down = 0.7,
  MAF_cutoff = 0.01,
  train = FALSE,
  cpu = 1,
  pop = "EAS"
)
}
\arguments{
\item{rgData}{Noob and dye-bias corrected signals produced by using \code{correct_noob_dye}.}

\item{plotBeta}{If TRUE, plot beta distributions for reference homozygous, heterozygous, and alternative homozygous.}

\item{vcf}{If TRUE, will write a VCF file in the current directory.}

\item{vcfName}{VCF file name. Only effective when vcf=TRUE.}

\item{R2_cutoff_up, R2_cutoff_down}{R-square cutoffs to filter variants (Variants with R-square > R2_cutoff_up or < R2_cutoff_down should be removed). Note that for VCF output, variants with R-square outside this range will be marked in the \code{FILTER} column. For the returned matrix, variants with R-square outside this range will be removed.}

\item{MAF_cutoff}{A MAF cutoff to filter variants. Note that for VCF output, variants with MAF below the cutoff will be marked in the \code{FILTER} column. For the returned matrix, variants with MAF below the cutoff will be removed.}

\item{train}{If TRUE, will fit the distribution of RAI (Ratio of Alternative allele Intensity) and filter probes by number of peaks. If FALSE, will use predefined probe list.}

\item{cpu}{Number of CPU. Only effective when train=TRUE.}

\item{pop}{Population. One of EAS, AMR, AFR, EUR, SAS, and ALL. Only probes with MAF of matching population > 0.01 will be kept. Only effective when train=TRUE.}
}
\value{
A matrix of genotype calls.
}
\description{
Call genotypes for Type I CCS probes
}
