% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgroundCorrection.R
\name{correct_noob_dye}
\alias{correct_noob_dye}
\title{Noob and dye-bias correction}
\usage{
correct_noob_dye(target, mnfst, cpu = 1)
}
\arguments{
\item{target}{A data frame with the}

\item{mnfst}{Manifest file. Required columns: Name, AddressA_ID (numeric), AddressB_ID (numeric), Infinium_Design_Type, and Color_Channel}

\item{cpu}{Number of CPU.}
}
\value{
A list of noob and dye-bias corrected signals containing:
\item{AR}{ - A matrix of probeA signals in Red channel}
\item{AG}{ - A matrix of probeA signals in Green channel}
\item{BR}{ - A matrix of probeB signals in Red channel}
\item{BG}{ - A matrix of probeB signals in Green channel}
}
\description{
Noob and dye-bias correction
}
