% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R,
%   R/data_representation_deprecated.R
\name{dosage2hard}
\alias{dosage2hard}
\title{Get hard genotypes from genotype probabilities}
\usage{
dosage2hard(genotypes)

dosage2hard(genotypes)
}
\arguments{
\item{genotypes}{Genotype calls.}
}
\value{
A matrix of hard genotypes.

A matrix of hard genotypes.
}
\description{
Hard genotype is defined as the genotype with highest genotype probability.

Hard genotype is defined as the genotype with highest genotype probability.
}
