% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes.R
\name{eBeta}
\alias{eBeta}
\title{Moments estimator for beta distribution (adapted from ewastools)}
\usage{
eBeta(x, w)
}
\arguments{
\item{x}{A vector of RAI values.}

\item{w}{Weights.}
}
\value{
A list of beta distribution shapes.
}
\description{
Moments estimator for beta distribution (adapted from ewastools)
}
