% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes_bayesian.R
\name{eBeta}
\alias{eBeta}
\title{Moments estimator for beta distribution (adapted from ewastools)}
\usage{
eBeta(x, w)
}
\description{
Moments estimator for beta distribution (adapted from ewastools)
}
