% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R
\name{filter_by_HWE}
\alias{filter_by_HWE}
\title{Filter by Hardy–Weinberg Equilibrium (HWE) p value}
\usage{
filter_by_HWE(hwe_p, hwe_p_cutoff)
}
\arguments{
\item{hwe_p}{Hardy–Weinberg Equilibrium (HWE) p values}

\item{hwe_p_cutoff}{A HWE p value cutoff to filter variants.}
}
\value{
Whether the variant passed filtering.
}
\description{
Variants with Hardy–Weinberg Equilibrium (HWE) p value < 1e-6 are recommended to remove.
}
