% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R
\name{filter_by_R2}
\alias{filter_by_R2}
\title{Filter by R2}
\usage{
filter_by_R2(R2, R2_cutoff_up = 1.1, R2_cutoff_down = 0.7)
}
\arguments{
\item{R2}{R-square}

\item{R2_cutoff_up}{Variants with R-square greater than this cutoff should be removed.}

\item{R2_cutoff_down}{Variants with R-square less than this cutoff should be removed.}
}
\value{
Whether the variant passed filtering.
}
\description{
Variants with R2>1.1 (marked as .) or R2<0.7 are recommended to remove.
}
