% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R
\name{format_genotypes}
\alias{format_genotypes}
\title{Format genotype calls}
\usage{
format_genotypes(
  genotypes,
  vcf = FALSE,
  vcfName,
  R2_cutoff_up = 1.1,
  R2_cutoff_down = 0.75,
  MAF_cutoff = 0.01,
  pop = "ALL",
  type,
  plotAF = FALSE,
  platform = "EPIC"
)
}
\arguments{
\item{genotypes}{Genotype calls.}

\item{vcf}{If TRUE, will write a VCF file in the current directory.}

\item{vcfName}{VCF file name. Only effective when vcf=TRUE.}

\item{R2_cutoff_up, R2_cutoff_down}{R-square cutoffs to filter variants (Variants with R-square > R2_cutoff_up or < R2_cutoff_down should be removed). Note that for VCF output, variants with R-square outside this range will be marked in the \code{FILTER} column. For the returned matrix, variants with R-square outside this range will be removed.}

\item{MAF_cutoff}{An MAF cutoff to filter variants. Note that for VCF output, variants with MAF below the cutoff will be marked in the \code{FILTER} column. For the returned matrix, variants with MAF below the cutoff will be removed.}

\item{pop}{Population to be used to extract AFs. One of EAS, AMR, AFR, EUR, SAS, and ALL.}

\item{type}{One of snp_probe, typeI_ccs_probe, and typeII_ccs_probe.}

\item{plotAF}{To plot the distribution of AFs in 1KGP and input data.}

\item{platform}{EPIC or 450K.}
}
\value{
A matrix of genotype calls. Variants with R2 or MAF beyond the cutoffs are removed. Genotypes with genotype quality (GQ) < 20 will be marked as NA.
}
\description{
Format genotype calls
}
