% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R
\name{format_genotypes}
\alias{format_genotypes}
\title{Format genotype calls produced by ewastools}
\usage{
format_genotypes(genotypes, vcf = FALSE, type)
}
\arguments{
\item{genotypes}{Genotype calls produced by using \code{ewastools::call_genotypes}.}

\item{vcf}{If TRUE, will write a VCF file in the current directory.}

\item{type}{One of genotyping_probe, ccs_snp_probe, and typeII_snp_probe.}
}
\value{
A matrix of genotype calls.
}
\description{
Format genotype calls produced by ewastools
}
