% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes_bayesian.R
\name{get_AF}
\alias{get_AF}
\title{Extract AFs from matching population in the 1000 Genomes Project (1KGP)}
\usage{
get_AF(RAI, pop = "EAS", type)
}
\arguments{
\item{RAI}{A matrix of RAI (Ratio of Alternative allele Intensity) for probes. Provide probes as rows and samples as columns.}

\item{pop}{Population to be used to extract AFs. One of EAS, AMR, AFR, EUR, SAS, and ALL.}

\item{type}{One of snp_probe, typeI_ccs_probe, and typeII_ccs_probe.}
}
\value{
A vector of AFs
}
\description{
Extract AFs from matching population in the 1000 Genomes Project (1KGP)
}
