% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes_bayesian.R
\name{get_GP}
\alias{get_GP}
\title{Infer posterior genotype probabilities based on the Bayesian approach}
\usage{
get_GP(RAI, weights, shapes, bayesian = TRUE, AF)
}
\arguments{
\item{RAI}{A MxN matrix of RAI (Ratio of Alternative allele Intensity). Provide probes as rows and samples as columns.}

\item{weights}{Prior probabilities that the RAI values belong to one of the three genotypes.}

\item{shapes}{A data frame (3x2) containing the two shapes for beta distributions of the three clusters.}

\item{bayesian}{Use the Bayesian approach or not.}

\item{AF}{A MxN matrix of AFs. Provide SNPs as rows and samples as columns. Only effective when bayesian=TRUE.}
}
\value{
A list containing
\item{pAA}{Posterior genotype probability of AA}
\item{pAB}{Posterior genotype probability of AB}
\item{pBB}{Posterior genotype probability of BB}
}
\description{
Prior genotype probabilities were inferred from AFs. The AFs can be in population level or individual-specific level. For population level AFs, they can be extracted from the matched population in the 1000 Genomes Project (1KGP). For individual-specific AFs, they can be calculated according to the top four PCs.
}
