% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/call_genotypes_bayesian.R
\name{get_PL}
\alias{get_PL}
\title{Calculate Phred-scaled genotype likelihood (PL)}
\usage{
get_PL(GP)
}
\arguments{
\item{GP}{A list of posterior genotype probabilities of genotype AA, AB, and BB.}
}
\value{
A list of Phred-scaled genotype likelihoods containing PL(AA), PL(AB), and PL(BB).
}
\description{
Calculate Phred-scaled genotype likelihood (PL)
}
