% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typeII_snp_probe.R
\name{getAB_typeII}
\alias{getAB_typeII}
\title{Get allelic balances for Type-II SNP probes}
\usage{
getAB_typeII(rgData)
}
\arguments{
\item{rgData}{Noob and dye-bias corrected signals produced by using \code{correct_noob_dye}.}
}
\value{
Allelic balances.
}
\description{
Get allelic balances for Type-II SNP probes
}
