% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R
\name{getHWE}
\alias{getHWE}
\title{Calculate Hardy-Weinberg Equilibrium (HWE) p value}
\usage{
getHWE(hardgeno)
}
\arguments{
\item{hardgeno}{A matrix of hard genotypes, with each row indicates a SNP and each column indicates a sample.}
}
\value{
HWE p values.
}
\description{
Calculate Hardy-Weinberg Equilibrium (HWE) p value
}
