% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_genotype_calling.R
\name{getKinship}
\alias{getKinship}
\title{Get kinship coefficients using the SEEKIN estimator}
\usage{
getKinship(dosage)
}
\arguments{
\item{dosage}{Dosage genotypes.}
}
\value{
A data frame containing kinship coefficient (Phi) and sample relationships between each two samples.
}
\description{
For homogeneous samples. Only SNPs with missing rate < 10\% were used.
}
