% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDetection.R
\name{getMod}
\alias{getMod}
\title{Filter probes according to the number and position of peaks}
\usage{
getMod(AB_geno, cpu = 1)
}
\arguments{
\item{AB_geno}{Matrix of allelic balances.}

\item{cpu}{Number of CPU.}
}
\value{
Probe filtering results
}
\description{
Filter probes according to the number and position of peaks
}
