% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakDetection.R
\name{getMod}
\alias{getMod}
\title{Filter probes according to the number and position of peaks}
\usage{
getMod(RAI, cpu = 1)
}
\arguments{
\item{RAI}{Matrix of RAI (Ratio of Alternative allele Intensity).}

\item{cpu}{Number of CPU.}
}
\value{
Probe filtering results
}
\description{
Filter probes according to the number and position of peaks
}
