% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_representation.R
\name{plot_beta_distribution}
\alias{plot_beta_distribution}
\title{Plot beta distributions for reference homozygous, heterozygous, and alternative homozygous}
\usage{
plot_beta_distribution(genotypes, type)
}
\arguments{
\item{genotypes}{Genotype calls produced by using \code{ewastools::call_genotypes}.}

\item{type}{One of "genotyping_probe", "ccs_snp_probe", and "typeII_snp_probe".}
}
\description{
Plot beta distributions for reference homozygous, heterozygous, and alternative homozygous
}
