% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probeInfo_typeII}
\alias{probeInfo_typeII}
\title{Type II CCS probe information for EPIC}
\format{
A data frame with 26420 rows and 16 columns:
\describe{
\item{Chr}{Chromosome ID}
\item{Pos}{Position}
\item{SNP}{SNP ID targeted by the CpG}
\item{RefAllele}{Reference allele}
\item{AltAllele}{Alternative allele}
\item{CpG}{CpG}
\item{Color}{Color channel}
\item{Group}{Probe types, color channel, and signal corresponds to alternative allele}
\item{ALL_AF}{Allele frequency of all population}
\item{EAS_AF}{Allele frequency of East Asian}
\item{AMR_AF}{Allele frequency of American}
\item{AFR_AF}{Allele frequency of African}
\item{EUR_AF}{Allele frequency of European}
\item{SAS_AF}{Allele frequency of South Asian}
\item{h_0.1}{Passed peak density test or not}
\item{loc_pass}{Passed peak position test or not}
}
}
\source{
\url{https://webdata.illumina.com/downloads/productfiles/methylationEPIC/infinium-methylationepic-v-1-0-b5-manifest-file-csv.zip}
}
\usage{
data(probeInfo_typeII)
}
\description{
A dataset containing information of Type II CCS probes with SNPs at the extension bases. We only consider the situation that the alternative allele is A/T and the reference allele is C/G.
}
\keyword{datasets}
