% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probelist}
\alias{probelist}
\title{Probe list}
\format{
A data frame with 27190 rows and 2 columns:
\describe{
\item{CpG}{CpG list}
\item{Type}{Probe types}
}
}
\usage{
data(probelist)
}
\description{
A dataset containing the list of 53 SNP probes on autosomes, 717 CCS SNP probes, and 27879 type-II probes.
}
\seealso{
Probe informations: \code{\link{probeInfo_snp()}}, \code{\link{probeInfo_ccs()}}, and \code{\link{probeInfo_typeII()}}.
}
\keyword{datasets}
