% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probelist}
\alias{probelist}
\title{Probe list for EPIC}
\format{
A data frame with 27188 rows and 2 columns:
\describe{
\item{CpG}{CpG list}
\item{Type}{Probe types}
}
}
\usage{
data(probelist)
}
\description{
A dataset containing the list of 53 SNP probes on autosomes, 715 Type I CCS probes, and 26420 type II CCS probes.
}
\seealso{
Probe informations: \code{\link{probeInfo_snp()}}, \code{\link{probeInfo_typeI()}}, and \code{\link{probeInfo_typeII()}}.
}
\keyword{datasets}
