% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{probelist}
\alias{probelist}
\title{Probe list}
\format{
A data frame with 5057 rows and 2 columns:
\describe{
\item{CpG}{CpG list}
\item{Type}{Probe types}
}
}
\usage{
data(probelist)
}
\description{
A dataset containing the list of 53 genotyping probes on autosomes, 132 CCS SNP probes, and 4872 type-II SNP probes.
}
\seealso{
Probe informations: \code{\link{probeInfo_geno()}}, \code{\link{probeInfo_ccs()}}, and \code{\link{probeInfo_typeII()}}.
}
\keyword{datasets}
