% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_individual_AF.R
\name{projection}
\alias{projection}
\title{PCA and Procrustes analysis}
\usage{
projection(studyGeno, plotPCA = TRUE, cpu = 1)
}
\arguments{
\item{studyGeno}{A matrix of genotypes of study samples. Provide probes as rows and samples as columns. Include all SNP probes, type I CCS probes, and type II CCS probes if available.}

\item{plotPCA}{To plot the projection of study samples in reference ancestry space.}

\item{cpu}{Number of CPU.}
}
\value{
A list containing
\item{refPC}{Top PCs in the reference}
\item{studyPC}{Top PCs in study samples}
}
\description{
PCA and Procrustes analysis
}
