% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recalibrate_genotypes.R
\name{recal_Geno}
\alias{recal_Geno}
\title{Recalibrate genotypes for samples of mixed population}
\usage{
recal_Geno(genotypes, type, refPC, studyPC)
}
\arguments{
\item{genotypes}{A list returned by either \code{callGeno_snp}, \code{callGeno_typeI}, or \code{callGeno_typeII} function.}

\item{type}{One of snp_probe, typeI_ccs_probe, and typeII_ccs_probe.}

\item{refPC}{Top PCs in the reference}

\item{studyPC}{Top PCs in study samples}
}
\value{
A list of recalibrated genotypes containing
\item{dosage}{A matrix of genotype calls}
\item{genotypes}{A list containing RAI, fits, and Genotype probabilities}
}
\description{
Recalibrate genotypes for samples of mixed population
}
