% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refGeno_1KGP3}
\alias{refGeno_1KGP3}
\title{Reference genotypes in the 1000 Genomes Project}
\format{
A matrix with 28,619 rows and 2504 columns:
\describe{
\item{Row}{SNPs overlapping the methylation probes}
\item{Column}{Samples}
}
}
\usage{
data(refGeno_1KGP3)
}
\description{
A matrix of reference genotypes in the 1000 Genomes Project (1KGP). It contains 2504 samples and 28,619 SNPs overlapping the methylation probes.
}
\keyword{datasets}
